+++ Asteroid_OtrantoDario/Asteroid_OtrantoDario/Objects/Button/button.cpp
@@ -3,11 +3,11 @@
 Button initButton(int initialPosX, int initialPosY, int fontSize, int buttonWidth, int buttonHeight, int optionNumber, string text, Color buttonColor, Color selectionColor)
 {
 	Button aux;
-	aux.rect.x = initialPosX;
-	aux.rect.y = initialPosY;
+	aux.rect.x = static_cast<float>(initialPosX);
+	aux.rect.y = static_cast<float>(initialPosY);
 	aux.fontSize = fontSize;
-	aux.rect.width = buttonWidth;
-	aux.rect.height = buttonHeight;
+	aux.rect.width = static_cast<float>(buttonWidth);
+	aux.rect.height = static_cast<float>(buttonHeight);
 	aux.optionNumber = optionNumber;
 	aux.text = text;
 	aux.buttonColor = buttonColor;
